examples/requirements.txt
django-jquery==1.9.1
django-autocomplete-light==1.4.9
